+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/internal/exportimport/data/handler/JournalPortletDataHandler.java
@@ -196,6 +196,7 @@ public class JournalPortletDataHandler extends BasePortletDataHandler {
 			new PortletDataHandlerBoolean(
 				NAMESPACE, "folders", true, false, null,
 				JournalFolder.class.getName()));
+		setStagingControls(getExportControls());
 	}
 
 	@Override
@@ -391,9 +392,15 @@ public class JournalPortletDataHandler extends BasePortletDataHandler {
 
 			_staging.populateLastPublishDateCounts(
 				portletDataContext,
-				new String[] {
-					DDMStructure.class.getName(), DDMTemplate.class.getName(),
-					JournalFeed.class.getName(), JournalFolder.class.getName()
+				new StagedModelType[] {
+					new StagedModelType(
+						DDMStructure.class.getName(),
+						JournalArticle.class.getName()),
+					new StagedModelType(
+						DDMTemplate.class.getName(),
+						DDMStructure.class.getName()),
+					new StagedModelType(JournalFeed.class.getName()),
+					new StagedModelType(JournalFolder.class.getName())
 				});
 
 			_populateJournalArticleLastPublishDateCounts(portletDataContext);